Cleaned up code, removed ability to manually drag/drop posts
While working on telegram stuff, I found that I had really messy code so I cleaned most of it up. Removed the abillity to manually drag/drop posts because it was buggy, and I feel like it'd be better to put my efforts elsewhere and possibly come back to it at a later date.