organized code into multiple files + small cleanups
Created a js subfolder that contains pretty much all of the javascript code used in this project, making it much easer to add in things and actually see what code IS there. Cleaned up the styling of the code too, moving HTML inline styles to the CSS file and making a consistent comment style throughout the files. Updated ReadMe to better reflect where the project is at this point.